Abstract out prov generation into a seperate class
